refactor(CQ2-4): split bangumi_data.py (1120 lines) into bangumi_data/ package
- Split BangumiData into 5 files using mixin pattern:
  - __init__.py: BangumiData facade + config initialization + singleton
  - _http.py: _BufferedResponse + _request_with_retry (HTTP layer)
  - cache.py: CacheMixin (download/cache/preload/stats/clear/force_update)
  - matching.py: MatchingMixin (find_bangumi_id + title matching helpers)
  - index.py: IndexMixin (TMDB mapping + title index)
- Re-export httpx/time/ijson/_request_with_retry in __init__.py for test
  patch compatibility (app.utils.bangumi_data.{httpx,time,ijson,
  _request_with_retry})
- Use function-level `from . import _request_with_retry` in cache.py
  methods so test patches on app.utils.bangumi_data._request_with_retry
  take effect (module-level import would bind a stale reference)
- All 1894 tests pass; ruff clean
